requirements.txt
# numpy==1.14.4
# PIL== 8.0.0
tensorflow==2.2.0
# keras==2.3.0
aiohttp==3.6.2
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
uvicorn==0.11.3
